pyproject.toml
[project]
name = "anilistshare-new"
version = "0.1.0"
description = "A fun way to store your anime watchlist"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.3",
    "django-allauth[mfa,socialaccount]>=65.15.0",
    "django-stubs>=6.0.1",
    "django-watchfiles>=1.4.0",
    "djangorestframework>=3.17.1",
    "jwt>=1.4.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "django-browser-reload>=1.21.0",
]
